feat(utoopack): support pass utoopack.root for symlink situation
