+++ docs/docs/docs/api/config.md
@@ -1557,8 +1557,6 @@ $ npm install @babel/runtime --save-dev
 - 类型: `Object`
 - 默认值: `{}`
 
-> 目前 utoopack 处于内测状态, 如果遇到问题，欢迎通过 issue 反馈~
-
 使用 Rust Bundler [utoopack](http://github.com/utooland/utoo) 用于提升构建性能。
 
 通过配置来启用这个能力，目前 utoopack 对框架中大部分功能已经支持。
