+++ packages/bundler-utoopack/src/util.ts
@@ -1,0 +1,17 @@
+import { chalk } from '@umijs/utils';
+
+export function getDevBanner(
+  protocol: string,
+  host?: string,
+  port?: number,
+  ip?: string,
+) {
+  const hostStr = host === '0.0.0.0' ? 'localhost' : host;
+  const messages = [];
+  messages.push('  App listening at:');
+  messages.push(
+    `  - Local:   ${chalk.cyan(`${protocol}//${hostStr}:${port}`)}`,
+  );
+  messages.push(`  - Network: ${chalk.cyan(`${protocol}//${ip}:${port}`)}`);
+  return messages.join('\n');
+}
