+++ packages/preset-umi/src/features/utoopack/utoopack.ts
@@ -11,7 +11,11 @@ export default (api: IApi) => {
     key: 'utoopack',
     config: {
       schema({ zod }) {
-        return zod.object({}).partial();
+        return zod
+          .object({
+            root: zod.string(),
+          })
+          .partial();
       },
     },
     enableBy: () =>
